changed float32 to float64 for height of user

+++ user/user.go
@@ -5,6 +5,6 @@ type User struct {
 	Fname   string
 	City    string
 	Phone   int64
-	Height  float32
+	Height  float64
 	Married bool
 }
